django/requirements.txt
aiohttp==3.9.1
colour-science==0.4.4 
cryptography==41.0.7
# daphne==4.0.0
dj-stripe==2.8.3
Django==5.0.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-filter==23.2
django-formtools==2.4.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-storages==1.13.2
django-timezone-field==5.0
django-versatileimagefield==3.0
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0 
google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
HAP-python==4.9.1
huesdk==1.7
ipython==8.20.0
mssql-django==1.3
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.4.4
zeroconf==0.131.0
psutil==5.9.8
scapy==2.5.0